frontend-plugin-api: refactor createDataContainer -> createExtensionDataContainer
