+++ packages/frontend-plugin-api/src/wiring/index.ts
@@ -17,7 +17,6 @@
 export { coreExtensionData } from './coreExtensionData';
 export {
   createExtension,
-  type ExtensionDataContainer,
   type ExtensionDefinition,
   type CreateExtensionOptions,
   type ExtensionDataValues,
@@ -32,6 +31,7 @@ export {
   type ExtensionInput,
   type LegacyExtensionInput,
 } from './createExtensionInput';
+export { type ExtensionDataContainer } from './createExtensionDataContainer';
 export {
   createExtensionDataRef,
   type AnyExtensionDataRef,
